Move test files and resolve issue with relative reference in schema
